requirements.txt
streamlit
langchain
langchain-community
langchain_groq
python-dotenv
scrapy
beautifulsoup4